pyproject.toml
[project]
name = "email-assistant-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "fastcrud>=0.15.7",
    "greenlet>=3.2.0",
    "openai>=1.75.0",
    "python-jose>=3.4.0",
    "sqlmodel>=0.0.24",
]
